chore(suite): decouple the desktop an mobile account menus
